# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 1 */
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  xp: number;
  /** Overall account level derived from XP */
  level: number;
  /** Current daily streak in days */
  streak: number;
  hasProgram: boolean;
  onboarded: boolean;
  createdAt?: string;
}

export interface AuthResult {
  user: User;
  /** Bearer token for mobile clients. Web uses the session cookie. */
  token: string;
}

export interface Profile {
  userId: number;
  onboarded: boolean;
  /**
     * What the user said about themselves
     * @nullable
     */
  about?: string | null;
  /** @nullable */
  currentSituation?: string | null;
  /** @nullable */
  biggestChallenge?: string | null;
  /** @nullable */
  motivation?: string | null;
  /** @nullable */
  focusMinutesPerDay?: number | null;
}

export interface ProfileUpdate {
  about?: string;
  currentSituation?: string;
  biggestChallenge?: string;
  motivation?: string;
  focusMinutesPerDay?: number;
  onboarded?: boolean;
}

export interface Path {
  /** One of millionaire, business, heartbreak, happiness, superhuman, confidence, focus */
  key: string;
  title: string;
  tagline: string;
  description: string;
  /** A hex color hint associated with the path */
  accent: string;
  outcomes?: string[];
}

export interface ProgramInput {
  pathKey: string;
  /** 30 or 60 */
  durationDays: number;
  about?: string;
  currentSituation?: string;
  biggestChallenge?: string;
  motivation?: string;
}

export interface Program {
  id: number;
  pathKey: string;
  pathTitle: string;
  title?: string;
  durationDays: number;
  /** active, completed, or archived */
  status: string;
  currentLevel: number;
  totalLevels: number;
  progressPct: number;
  /** @nullable */
  accent?: string | null;
  createdAt: string;
}

export interface Task {
  id: number;
  levelId: number;
  levelNumber: number;
  dayNumber: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /**
     * Suggested time like "07:00" or "Morning"
     * @nullable
     */
  timeOfDay?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  completed: boolean;
  /** @nullable */
  completedAt?: string | null;
  xp: number;
}

export interface Level {
  id: number;
  levelNumber: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** locked, active, or completed */
  status: string;
  xpReward: number;
  tasks: Task[];
}

export interface ProgramDetail {
  id: number;
  pathKey: string;
  pathTitle: string;
  title?: string;
  /** @nullable */
  summary?: string | null;
  durationDays: number;
  status: string;
  currentLevel: number;
  totalLevels: number;
  progressPct: number;
  /** @nullable */
  accent?: string | null;
  createdAt: string;
  levels: Level[];
}

export interface Badge {
  key: string;
  title: string;
  description: string;
  /** @nullable */
  icon?: string | null;
  earned: boolean;
  /** @nullable */
  earnedAt?: string | null;
}

export interface TaskCompletionResult {
  task: Task;
  xpAwarded: number;
  totalXp: number;
  streak: number;
  leveledUp: boolean;
  newBadges?: Badge[];
}

export interface DashboardSummary {
  hasProgram: boolean;
  program?: Program;
  xp: number;
  level: number;
  streak: number;
  longestStreak?: number;
  todayTotal: number;
  todayCompleted: number;
  todayTasks?: Task[];
  /** @nullable */
  nextLevelTitle?: string | null;
  /**
     * A motivational line for the day
     * @nullable
     */
  quote?: string | null;
}

export interface DayActivity {
  date: string;
  completed: number;
}

export interface Progress {
  xp: number;
  level: number;
  streak: number;
  longestStreak: number;
  totalTasksCompleted: number;
  completionRate: number;
  weeklyActivity: DayActivity[];
  badges: Badge[];
}

export interface Reminder {
  id: number;
  title: string;
  /** HH:MM 24-hour */
  timeOfDay: string;
  enabled: boolean;
  /** 0=Sunday .. 6=Saturday */
  daysOfWeek: number[];
}

export interface ReminderInput {
  title: string;
  timeOfDay: string;
  enabled?: boolean;
  daysOfWeek?: number[];
}

export interface ReminderUpdate {
  title?: string;
  timeOfDay?: string;
  enabled?: boolean;
  daysOfWeek?: number[];
}

export interface ChatMessage {
  id: number;
  /** user or assistant */
  role: string;
  content: string;
  createdAt: string;
}

export interface ChatMessageInput {
  /** @minLength 1 */
  content: string;
}

export interface ChatReply {
  userMessage: ChatMessage;
  assistantMessage: ChatMessage;
}