# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task complete and award XP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCompletionResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{id}/uncomplete:
    post:
      operationId: uncompleteTask
      tags: [tasks]
      summary: Mark a task not complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Task:
      type: object
      required: [id, levelId, levelNumber, dayNumber, title, completed, xp]
      properties:
        id:
          type: integer
        levelId:
          type: integer
        levelNumber:
          type: integer
        dayNumber:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        timeOfDay:
          type: ["string", "null"]
          description: Suggested time like "07:00" or "Morning"
        durationMinutes:
          type: ["integer", "null"]
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
        xp:
          type: integer
    TaskCompletionResult:
      type: object
      required: [task, xpAwarded, totalXp, streak, leveledUp]
      properties:
        task:
          $ref: "#/components/schemas/Task"
        xpAwarded:
          type: integer
        totalXp:
          type: integer
        streak:
          type: integer
        leveledUp:
          type: boolean
        newBadges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
...
    Badge:
      type: object
      required: [key, title, description, earned]
      properties:
        key:
          type: string
        title:
          type: string
        description:
          type: string
        icon:
          type: ["string", "null"]
        earned:
          type: boolean
        earnedAt:
          type: ["string", "null"]
          format: date-time
...